Fetch backend for Scala.js
Create a JavaScript backend implementation based on the fetch API. The
fetch API does not allow access to redirect responses so the backend
cannot be wrapped with the `FollowRedirectsBackend`.

In addition a streaming implentation using Monix has also been created.
Unfortunately no browsers support a stream as a request body so it is
converted into an in memory array before sending.