Filter out content header when redirecting to GET
In #182, the possibility to redirect POST and PUT to a GET has been
introduced. When redirecting the content is dropped. However, if the
content headers are not cleaned, in particuler the `Content-Length` one,
the underlying backend might hang, waiting for body data to be
available. Because no content is sent in the redirect, it is safe and
better to remove all content related headers in the redirected request.