Fix handling of utf-8 responses in scala-native
Problem:
`str.toCharArray.map(_.toByte)`, takes only one byte per
character, thus producing an incorrect `Array[Byte]` for
any string that contains multi-byte characters.

Solution:
Use `str.getBytes` instead.

Testing:
This patch also adds an unicode character to the test
string in `SyncHttpTest`. The modified test fails without
the patch, and passes with the patch.

Future Work:
Currently `AbstractCurlBackend` calls the `fromCString`
function from scala-native, without the encoding parameter.
Thus, the encoding parameter gets defaulted to UTF-8.
Testing needs to be done in cases where the encoding
of the server response is not UTF-8, which is out of the
scope for this patch.