Use scala-js-dom `File` type
Creating a `SttpFile` from a JavaScript file currently requires a custom
`File` type. This type exists as the one provided by scala-js-dom does
not provide the ability to create a file manually. However this makes it
annoying to use when you have a scala-js-dom `File`.

To improve this, the custom `File` type now extends the scala-js-dom
`File` and a `SttpFile` is now created from a scala-js-dom `File`.