test: redirection cookie behavior ignoring empty sensitive headers set
Redirects do not apply set-cookie in the request chain meaning
multistep request chains that track state will be missing cookies in the
middle of redirection, only real alternative is manual control

After redirection chain completes any set-cookies that started or were
in the middle of the request chain get ignored. Only the last responses set-cookie's
will be applied. It is possible to collect the cookies / headers from
response.history, but adds complexity.